back-end/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
boto3==1.26.114
botocore==1.29.114
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-redis==5.2.0
google-api-core==2.11.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
mysqlclient==2.1.1
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
quick-mailer==2022.2.22
redis==4.5.4
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
